requirements.txt
# QMail - Quantum-Secure Email Client Dependencies

# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Cryptography
cryptography==41.0.7
pycryptodome==3.19.0

# Post-Quantum Cryptography (if available, otherwise optional)
# pqcrypto==0.3.0  # Uncomment if PQC is needed

# HTTP Client for QKD API
requests==2.31.0

# Email Libraries (built-in, no need to install)
# smtplib, imaplib, email

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter for Railway

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# WSGI Server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# HTML Sanitization
bleach==6.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
